requirements.txt
django==4.2
djangorestframework==3.14
django-cors-headers
drf-spectacular
djangorestframework-simplejwt[crypto]==5.3.1
dj-rest-auth 
django-allauth
django-unfold